requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
jinja2==3.1.4
httpx==0.28.1
cryptography==44.0.0
slowapi==0.1.9
python-dotenv==1.0.1
pydantic-settings==2.7.0
